Move private helper definitions to nnf.util
